[Keychron C1 RGB] Move to rgb subfolder (#122)
* Move keychron c1 to c1/rgb

* Updated to the new path and board name
