+++ keyboards/keychron/c1/rgb/info.json
