+++ keyboards/keychron/c1/rgb/readme.md
@@ -10,7 +10,7 @@ A customizable TKL keyboard.
 
 Make example for this keyboard (after setting up your build environment):
 
-    make keychron/c1:default
+    make keychron/c1/rgb:default
 Flashing example for this keyboard:
 1. If your keyboard currently has stock firmware installed, put your keyboard first into bootloader by shorting the boot pins found under the spacebar before plugging in your keyboard to the PC. Otherwise, press `Fn + Esc` to put your keyboard into bootloader.
 1. Download and run [Sonix Flasher](https://github.com/SonixQMK/sonix-flasher/releases/tag/v0.2.1).
