+++ keyboards/keychron/c1/rgb/rgb.c
@@ -14,7 +14,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-#include "c1.h"
+#include "rgb.h"
 
 // Manage Windows and Mac LEDs
 // - Show status of mode switch
