+++ keyboards/keychron/c1/rgb/rgb.json
